# PERF: Optimize TsonConvert.cs