Hi @miou-gh, I've tried to make some performance optimizations to this repo. Following are the benchmark results before and after the changes:

**Benchmarks before changes:**
|              Method |         Mean |       Error |      StdDev |      Gen 0 | Gen 1 | Gen 2 |  Allocated |
|-------------------- |-------------:|------------:|------------:|-----------:|------:|------:|-----------:|
| DeserializeFromJSON |     326.1 us |     6.47 us |    10.63 us |    86.4258 |     - |     - |  176.95 KB |
| DeserializeFromTSON | 123,766.8 us | 2,466.17 us | 7,115.47 us | 10000.0000 |     - |     - | 21240.2 KB |
|     SerializeToJSON |     201.7 us |     1.62 us |     1.27 us |    62.7441 |     - |     - |  128.52 KB |
|     SerializeToTSON |   7,623.0 us |   152.12 us |   281.97 us |  1187.5000 |     - |     - | 2437.67 KB |

**Benchmarks after changes:**
|              Method |         Mean |       Error |      StdDev |      Gen 0 | Gen 1 | Gen 2 |   Allocated |
|-------------------- |-------------:|------------:|------------:|-----------:|------:|------:|------------:|
| DeserializeFromJSON |     318.5 us |     2.04 us |     1.71 us |    86.4258 |     - |     - |   176.95 KB |
| DeserializeFromTSON | 128,648.8 us | 2,543.90 us | 6,566.63 us | 10000.0000 |     - |     - | 21345.45 KB |
|     SerializeToJSON |     212.0 us |     4.03 us |     6.84 us |    62.7441 |     - |     - |   128.52 KB |
|     SerializeToTSON |   6,978.6 us |   134.88 us |   144.32 us |  1156.2500 |     - |     - |  2388.41 KB |

As you can see the allocations have gone down. I've also made sure that all the unit tests pass following these changes as well. Could you please confirm whether or not you think my changes are valid?

Thank you!